+++ src/components/dashboard/AddPlant/AddPlant.tsx
@@ -3,6 +3,7 @@ import "./AddPlant.scss";
 import { database } from "../../../firebase";
 import Modal from "../Modal/Modal";
 import { useAuth } from "../../../contexts/AuthContext";
+import AddPlantForm from "../AddPlantForm/AddPlantForm";
 
 export default function AddPlant() {
   const [modal, setModal] = useState(false);
@@ -47,11 +48,13 @@ export default function AddPlant() {
         Add Plant
       </button>
       {modal && (
-        <Modal
-          close={handleClose}
-          handleForm={handleForm}
-          handleSubmit={handleSubmit}
-        />
+        <Modal>
+          <AddPlantForm
+            close={handleClose}
+            handleForm={handleForm}
+            handleSubmit={handleSubmit}
+          />
+        </Modal>
       )}
     </>
   );
